PlantGeom

Documentation for PlantGeom.

PlantGeom.Material

PlantGeom.RefMesh

PlantGeom.RefMeshes

MakieCore.plot!

Meshes.nelements

Meshes.nvertices

PlantGeom.align_ref_meshes

PlantGeom.get_ref_meshes

PlantGeom.materialBDD_to_material

PlantGeom.meshBDD_to_meshes

PlantGeom.ref_meshes_color

PlantGeom.Material — Type

Data structure for a mesh material that is used to describe the light components of a Phong reflection type model. All data is stored as RGBα for Red, Green, Blue and transparency.

PlantGeom.RefMesh — Type

RefMesh type. Stores all information about a reference Mesh

PlantGeom.RefMeshes — Type

RefMeshes type. Stores all RefMesh.

MakieCore.plot! — Method

using MultiScaleTreeGraph, PlantGeom, WGLMakie, Meshes

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simpleOPFshapes.opf") opf = readopf(file) meshes = getref_meshes(opf)

viz(meshes)

With one shared color:

viz(meshes, color = :green)

One color per reference mesh:

viz(meshes, color = Dict(0 => :burlywood4, 1 => :springgreen4, 2 => :burlywood4))

Or just changing the color of some:

viz(meshes, color = Dict(0 => :burlywood4, 2 => :burlywood4))

One color for each vertex of the refmesh 0:

nvertices(meshes)[1] viz(meshes, color = Dict(0 => 1:nvertices(meshes)[0]))

Meshes.nelements — Method

nelements(meshes::RefMeshes)

Return the number of elements for each reference mesh as a dict of id => nelements

Meshes.nvertices — Method

nvertices(meshes::RefMeshes)

Return the number of vertices for each reference mesh as a dict of id => nvertices

PlantGeom.align_ref_meshes — Method

align_ref_meshes(meshes::RefMeshes)

Align all reference meshes along the X axis. Used for visualisation only.

PlantGeom.get_ref_meshes — Method

get_ref_meshes(mtg)

Get all reference meshes from an mtg, usually from an OPF.

Examples

using MultiScaleTreeGraph, PlantGeom
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshes = get_ref_meshes(opf)

using MeshViz
viz(meshes[0]["mesh"])
viz(meshes)
meshes.meshes

PlantGeom.materialBDD_to_material — Method

Parse a material in opf format to a material

PlantGeom.meshBDD_to_meshes — Method

meshBDD_to_meshes(x)

Examples

using MultiScaleTreeGraph
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshBDD_to_meshes(opf.attributes[:meshBDD])

PlantGeom.ref_meshes_color — Method

Get the reference meshes colors (only the diffuse part for now).

Examples

using MultiScaleTreeGraph, PlantGeom
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshes = get_ref_meshes(opf)
PlantGeom.ref_meshes_color(meshes)