PlantGeom

Documentation for PlantGeom.

PlantGeom.Material

PlantGeom.RefMesh

PlantGeom.RefMeshes

MakieCore.plot!

PlantGeom.align_ref_meshes

PlantGeom.get_ref_meshes

PlantGeom.materialBDD_to_material

PlantGeom.meshBDD_to_meshes

PlantGeom.meshes_color

PlantGeom.Material — Type

Data structure for a mesh material that is used to describe the light components of a Phong reflection type model. All data is stored as RGBα for Red, Green, Blue and transparency.

PlantGeom.RefMesh — Type

RefMesh type. Stores all information about a reference Mesh

PlantGeom.RefMeshes — Type

RefMeshes type. Stores all RefMesh.

MakieCore.plot! — Method

using MultiScaleTreeGraph, PlantGeom, WGLMakie

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simpleOPFshapes.opf") opf = readopf(file) meshes = getref_meshes(opf)

viz(meshes)

PlantGeom.align_ref_meshes — Method

align_ref_meshes(meshes::RefMeshes)

Align all reference meshes along the X axis. Used for visualisation only.

PlantGeom.get_ref_meshes — Method

get_ref_meshes(mtg)

Get all reference meshes from an mtg, usually from an OPF.

Examples

using MultiScaleTreeGraph, PlantGeom
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshes = get_ref_meshes(opf)

using MeshViz
viz(meshes[0]["mesh"])
viz(meshes)
meshes.meshes

PlantGeom.materialBDD_to_material — Method

Parse a material in opf format to a material

PlantGeom.meshBDD_to_meshes — Method

meshBDD_to_meshes(x)

Examples

using MultiScaleTreeGraph
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshBDD_to_meshes(opf.attributes[:meshBDD])

PlantGeom.meshes_color — Function

Find each mesh color.

Notes

meshes is typically the output from [meshes_from_topology].

Exemples

opf= readopf("simpleOPFshapes.opf") meshdata= meshesfromtopology(opf) shapes= extractopfshapes(opf) materials= extractopfmaterials(opf) cols= meshescolor(opf,meshdata) mesh(mergemeshes(meshdata["mesh"]), color= cols)