PlantGeom

Documentation for PlantGeom.

PlantGeom.Material

PlantGeom.RefMesh

PlantGeom.RefMeshes

MakieCore.plot!

MakieCore.plot!

Meshes.nelements

Meshes.nvertices

PlantGeom.align_ref_meshes

PlantGeom.get_attr_type

PlantGeom.get_ref_meshes

PlantGeom.get_ref_meshes_color

PlantGeom.materialBDD_to_material

PlantGeom.meshBDD_to_meshes

PlantGeom.parse_geometry

PlantGeom.parse_meshBDD!

PlantGeom.parse_opf_array

PlantGeom.parse_opf_attributeBDD!

PlantGeom.parse_opf_elements!

PlantGeom.parse_opf_topology!

PlantGeom.parse_ref_meshes

PlantGeom.read_opf

PlantGeom.refmesh_to_mesh

PlantGeom.taper

PlantGeom.Material — Type

Data structure for a mesh material that is used to describe the light components of a Phong reflection type model. All data is stored as RGBα for Red, Green, Blue and transparency.

PlantGeom.RefMesh — Type

RefMesh type. Stores all information about a reference Mesh

PlantGeom.RefMeshes — Type

RefMeshes type. Stores all RefMesh.

MakieCore.plot! — Method

using MultiScaleTreeGraph, PlantGeom, WGLMakie

file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simpleOPFshapes.opf")

file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","coffee.opf")

file = "D:/OneDrive - cirad.fr/TravailAMAP/Processes/LightinterceptionGPU/Julia3D/P6Ruii_L2P02.opf"

opf = readopf(file) refmeshes = getrefmeshes(opf)

viz(ref_meshes)

transform!(opf, (node -> refmeshtomesh(node, ref_meshes)) => :mesh)

viz(opf, color = Dict(0 => :burlywood4, 1 => :springgreen4))

With one shared color:

viz(meshes, color = :green)

One color per reference mesh:

viz(meshes, color = Dict(0 => :burlywood4, 1 => :springgreen4, 2 => :burlywood4))

Or just changing the color of some:

viz(meshes, color = Dict(0 => :burlywood4, 2 => :burlywood4))

One color for each vertex of the refmesh 0:

nvertices(meshes)[1] viz(meshes, color = Dict(0 => 1:nvertices(meshes)[0]))

MakieCore.plot! — Method

using MultiScaleTreeGraph, PlantGeom, WGLMakie, Meshes

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simpleOPFshapes.opf") opf = readopf(file) meshes = getref_meshes(opf)

viz(meshes)

With one shared color:

viz(meshes, color = :green)

One color per reference mesh:

viz(meshes, color = Dict(0 => :burlywood4, 1 => :springgreen4, 2 => :burlywood4))

Or just changing the color of some:

viz(meshes, color = Dict(0 => :burlywood4, 2 => :burlywood4))

One color for each vertex of the refmesh 0:

nvertices(meshes)[1] viz(meshes, color = Dict(0 => 1:nvertices(meshes)[0]))

Meshes.nelements — Method

nelements(meshes::RefMeshes)

Return the number of elements for each reference mesh as a dict of id => nelements

Meshes.nvertices — Method

nvertices(meshes::RefMeshes)

Return the number of vertices for each reference mesh as a dict of id => nvertices

PlantGeom.align_ref_meshes — Method

align_ref_meshes(meshes::RefMeshes)

Align all reference meshes along the X axis. Used for visualisation only.

PlantGeom.get_attr_type — Method

Get the attributes types in Julia DataType.

PlantGeom.get_ref_meshes — Method

get_ref_meshes(mtg)

Get all reference meshes from an mtg, usually from an OPF.

Examples

using MultiScaleTreeGraph, PlantGeom
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshes = get_ref_meshes(opf)

using MeshViz, GLMakie
viz(meshes)
meshes.meshes[0].material

PlantGeom.get_ref_meshes_color — Method

get_ref_meshes_color(meshes::RefMeshes)

Get the reference meshes colors (only the diffuse part for now).

Examples

using MultiScaleTreeGraph, PlantGeom
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshes = get_ref_meshes(opf)
PlantGeom.get_ref_meshes_color(meshes)

PlantGeom.materialBDD_to_material — Method

Parse a material in opf format to a material

PlantGeom.meshBDD_to_meshes — Method

meshBDD_to_meshes(x)

Examples

using MultiScaleTreeGraph
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshBDD_to_meshes(opf.attributes[:meshBDD])

PlantGeom.parse_geometry — Method

Parse the geometry element of the OPF.

Note

The transformation matrix is 3*4. elem = elem.content

PlantGeom.parse_meshBDD! — Method

Parse the meshBDD using [`parseopf_array`](@ref)

PlantGeom.parse_opf_array — Function

Parse an array of values from the OPF into a Julia array (Arrays in OPFs are not following XML recommendations)

PlantGeom.parse_opf_attributeBDD! — Method

Parse the opf attributes as a Dict.

PlantGeom.parse_opf_elements! — Method

Generic parser for OPF elements.

Arguments

opf::OrderedDict: the opf Dict (using [XMLDict.xml_dict])

elem_types::Array: the target types of the element (e.g. "[String, Int64]")

Details

elem_types should be of the same length as the number of elements found in each item of the subchild. elem_types = [Float64, Float64, Float64, Float64, Float64, Float64]

PlantGeom.parse_opf_topology! — Function

Parser for OPF topology.

Note

The transformation matrices in geometry are 3*4. parseopftopology!(elem, nodei, features) node = elem mtg = nodei features = getattrtype(opf_attr[:attributeBDD])

Debugging:

mtg = nothing

node = elem mtg = nodei parseopftopology!( node, nothing, getattrtype(opfattr[:attributeBDD]), attrtype, mtgtype )

PlantGeom.parse_ref_meshes — Method

parse_ref_meshes(mtg)

Parse the reference meshes of an OPF into RefMeshes.

PlantGeom.read_opf — Function

read_opf(file, attr_type = Dict, mtg_type = MutableNodeMTG)

Read an OPF file, and returns an MTG.

Arguments

file::String: The path to the opf file.

attr_type::DataType = Dict: the type used to hold the attribute values for each node.

mtg_type = MutableNodeMTG: the type used to hold the mtg encoding for each node (i.e.

link, symbol, index, scale). See details section below.

Details

attr_type should be:

NamedTuple if you don't plan to modify the attributes of the mtg, e.g. to use them for

plotting or computing statistics...

MutableNamedTuple if you plan to modify the attributes values but not adding new attributes

very often, e.g. recompute an attribute value...

Dict or similar (e.g. OrderedDict) if you plan to heavily modify the attributes, e.g.

adding/removing attributes a lot

The MultiScaleTreeGraph package provides two types for mtg_type, one immutable (NodeMTG), and one mutable (MutableNodeMTG). If you're planning on modifying the mtg encoding of some of your nodes, you should use MutableNodeMTG, and if you don't want to modify anything, use NodeMTG instead as it should be faster.

Note

See the documentation of the MTG format from the package documentation for further details, e.g. The MTG concept.

Returns

The MTG root node.

Examples

using PlantGeom
file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)

PlantGeom.refmesh_to_mesh — Method

refmesh_to_mesh(node, ref_meshes)

Compute a node mesh based on the reference mesh, the transformation matrix and the tapering.

Examples

using PlantGeom
file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)

node = opf[1][1][1]

new_mesh = refmesh_to_mesh(node, get_ref_meshes(opf))

using MeshViz, GLMakie
viz(new_mesh)

PlantGeom.taper — Method

Returns a tapered mesh using dDwn and dUp based on the geometry of an input mesh. Tapering a mesh transforms it into a tapered version (i.e. pointy) or enlarged object, e.g. make a cone from a cylinder.