API

PlantGeom.Material

PlantGeom.Phong

PlantGeom.RefMesh

PlantGeom.RefMeshes

PlantGeom.geometry

MeshViz.viz

MeshViz.viz!

Meshes.nelements

Meshes.nelements

Meshes.nvertices

Meshes.nvertices

PlantGeom.align_ref_meshes

PlantGeom.attributes_to_xml

PlantGeom.coordinates!

PlantGeom.cylinder

PlantGeom.diagram

PlantGeom.diagram!

PlantGeom.extend_pos

PlantGeom.get_attr_type

PlantGeom.get_color

PlantGeom.get_colormap

PlantGeom.get_ref_mesh_index

PlantGeom.get_ref_mesh_index!

PlantGeom.get_ref_meshes

PlantGeom.get_ref_meshes_color

PlantGeom.map_coord

PlantGeom.materialBDD_to_material

PlantGeom.meshBDD_to_meshes

PlantGeom.mtg_coordinates_df

PlantGeom.mtg_to_opf_link

PlantGeom.mtg_topology_to_xml!

PlantGeom.parse_geometry

PlantGeom.parse_materialBDD!

PlantGeom.parse_meshBDD!

PlantGeom.parse_opf_array

PlantGeom.parse_opf_attributeBDD!

PlantGeom.parse_opf_elements!

PlantGeom.parse_opf_topology!

PlantGeom.parse_ref_meshes

PlantGeom.read_opf

PlantGeom.read_ply

PlantGeom.refmesh_to_mesh

PlantGeom.refmesh_to_mesh!

PlantGeom.rotate_point

PlantGeom.taper

PlantGeom.transform_mesh!

PlantGeom.write_opf

PlantGeom.xmax

PlantGeom.xmin

PlantGeom.ymax

PlantGeom.ymin

PlantGeom.zmax

PlantGeom.zmin

RecipesBase.plot

RecipesBase.plot!

PlantGeom.Material — Type

A material for the illumination model (e.g. Phong illumination).

PlantGeom.Phong — Type

Data structure for a mesh material that is used to describe the light components of a Phong reflection type model. All data is stored as RGBα for Red, Green, Blue and transparency.

PlantGeom.RefMesh — Type

RefMesh(
    name::S
    mesh::SimpleMesh
    normals::N
    texture_coords::T
    material::M
    taper::Bool
)

RefMesh(name, mesh, material = RGB(220 / 255, 220 / 255, 220 / 255))

RefMesh type. Stores all information about a Mesh:

name::S: the mesh name

mesh::SimpleMesh: the actual mesh information -> points and topology

normals::Vector{Float64}: the normals, given as a vector of x1,y1,z1,x2,y2,z2...

texture_coords::Vector{Float64}: the texture coordinates (not used yet), idem, a vector

material::M: the material, used to set the shading

taper::Bool: true if tapering is enabled

The reference meshes are then transformed on each node of the MTG using a transformation matrix to match the actual mesh.

PlantGeom.RefMeshes — Type

RefMeshes type. Data base that stores all RefMesh in an MTG. Usually stored in the :ref_meshes attribute of the root node.

PlantGeom.geometry — Type

geometry(
    ref_mesh::M
    ref_mesh_index::Union{Int,Nothing}
    transformation::T
    dUp::S
    dDwn::S
    mesh::Union{SimpleMesh,Nothing}
)

A Node geometry with the reference mesh, its transformation matrix and optionnally the index of the reference mesh in the reference meshes data base (see notes) and the resulting mesh (optional to save memory).

Note

The refmesh usually points to a RefMesh stored in the `:refmeshes` attribute of the root node of the MTG.

Although optinal, storing the index of the reference mesh (ref_mesh_index) in the database allows a faster writing of the MTG as an OPF to disk.

If no transformation matrix is needed, you can use I from the Linear Algebra package (lazy)

The transformation field should a CoordinateTransformations.jl's transformation. In case no transformation is needed, use IdentityTransformation(). If you already have the transformation matrix, you can pass it to LinearMap().

MeshViz.viz — Function

viz(opf::MultiScaleTreeGraph.Node; kwargs...)
viz!(opf::MultiScaleTreeGraph.Node; kwargs...)

Vizualise the 3D geometry of an MTG (usually read from an OPF). This function search for the :geometry attribute in each node of the MTG, and build the vizualisation using the mesh field, or the reference meshes and the associated transformation matrix if missing.

This function needs 3D information first.

Examples

using MultiScaleTreeGraph, PlantGeom, GLMakie

file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
# file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","coffee.opf")

opf = read_opf(file)
viz(opf)

# If you need to plot the opf several times, you better cache the mesh in the node geometry
# like so:
transform!(opf, refmesh_to_mesh!)

# Then plot it again like before, and it will be faster:
viz(opf)

# We can also color the 3d plot with several options:
# With one shared color:
viz(opf, color = :red)
# One color per reference mesh:
viz(opf, color = Dict(1 => :burlywood4, 2 => :springgreen4, 3 => :burlywood4))

# Or just changing the color of some:
viz(opf, color = Dict(1 => :burlywood4))
# One color for each vertex of the refmesh 1:
viz(opf, color = Dict(1 => 1:nvertices(get_ref_meshes(opf))[1]))

# Or coloring by opf attribute, e.g. using the mesh max Z coordinates (NB: need to use
# `refmesh_to_mesh!` before, see above):
transform!(opf, :geometry => (x -> zmax(x.mesh)) => :z_max, ignore_nothing = true)
viz(opf, color = :z_max)

# Or even coloring by the value of the Z coordinates of each vertex:
transform!(opf, :geometry => (x -> [i.coords[3] for i in x.mesh.points]) => :z, ignore_nothing = true)
viz(opf, color = :z, showfacets = true)

MeshViz.viz! — Function

viz(opf::MultiScaleTreeGraph.Node; kwargs...)
viz!(opf::MultiScaleTreeGraph.Node; kwargs...)

Vizualise the 3D geometry of an MTG (usually read from an OPF). This function search for the :geometry attribute in each node of the MTG, and build the vizualisation using the mesh field, or the reference meshes and the associated transformation matrix if missing.

This function needs 3D information first.

Examples

using MultiScaleTreeGraph, PlantGeom, GLMakie

file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
# file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","coffee.opf")

opf = read_opf(file)
viz(opf)

# If you need to plot the opf several times, you better cache the mesh in the node geometry
# like so:
transform!(opf, refmesh_to_mesh!)

# Then plot it again like before, and it will be faster:
viz(opf)

# We can also color the 3d plot with several options:
# With one shared color:
viz(opf, color = :red)
# One color per reference mesh:
viz(opf, color = Dict(1 => :burlywood4, 2 => :springgreen4, 3 => :burlywood4))

# Or just changing the color of some:
viz(opf, color = Dict(1 => :burlywood4))
# One color for each vertex of the refmesh 1:
viz(opf, color = Dict(1 => 1:nvertices(get_ref_meshes(opf))[1]))

# Or coloring by opf attribute, e.g. using the mesh max Z coordinates (NB: need to use
# `refmesh_to_mesh!` before, see above):
transform!(opf, :geometry => (x -> zmax(x.mesh)) => :z_max, ignore_nothing = true)
viz(opf, color = :z_max)

# Or even coloring by the value of the Z coordinates of each vertex:
transform!(opf, :geometry => (x -> [i.coords[3] for i in x.mesh.points]) => :z, ignore_nothing = true)
viz(opf, color = :z, showfacets = true)

Meshes.nelements — Method

nelements(meshes::RefMeshes)

Return the number of elements for each reference mesh as a vector of nelements

Meshes.nelements — Method

nelements(meshes::RefMeshes)

Return the number of elements of a reference mesh

Meshes.nvertices — Method

nvertices(meshes::RefMeshes)

Return the number of vertices for each reference mesh as a vector of nvertices

Meshes.nvertices — Method

nvertices(meshes::RefMesh)

Return the number of vertices of a reference mesh

PlantGeom.align_ref_meshes — Method

align_ref_meshes(meshes::RefMeshes)

Align all reference meshes along the X axis. Used for visualisation only.

PlantGeom.attributes_to_xml — Method

attributes_to_xml(node, xml_parent)

Write an MTG node into an XML node.

PlantGeom.coordinates! — Method

coordinates!(mtg; angle = 45; force = false)

Compute dummy 3d coordinates for the mtg nodes using an alterning phyllotaxy. Used when coordinates are missing. Coordinates are just node attributes with reserved names: :XX, :YY and :ZZ.

Returns

Nothing, mutates the mtg in-place (adds :XX, :YY and :ZZ to nodes).

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
coordinates!(mtg)
DataFrame(mtg, [:XX, :YY, :ZZ])

PlantGeom.cylinder — Method

cylinder()

Return a reference cylinder (RefMesh) of 1m radius and 1m length. The cylinder axis is oriented parallel to the z axis, and the origin (center of the bottom cap of the cylinder) is positionned at x = 0, y = 0 and z = 0.

PlantGeom.diagram — Function

diagram(opf::MultiScaleTreeGraph.Node; kwargs...)
diagram!(opf::MultiScaleTreeGraph.Node; kwargs...)

Make a diagram of the MTG tree, paired with a Makie.jl backend.

See also apply_recipe for the same plot with a Plots.jl backend.

The main attributes are:

color: the color of the nodes

colormap: the colormap used if the color uses an attribute. By default it uses viridis.

Must be a ColorScheme from ColorSchemes or a Symbol with its name.

Examples

using PlantGeom, GLMakie

file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
# file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","coffee.opf")

opf = read_opf(file)

diagram(opf)

# We can also color the 3d plot with several options:
# With one shared color:
diagram(opf, color = :red)

# Or colouring by opf attribute, *e.g.* using the nodes Z coordinates:
diagram(opf, color = :ZZ)

PlantGeom.diagram! — Function

diagram(opf::MultiScaleTreeGraph.Node; kwargs...)
diagram!(opf::MultiScaleTreeGraph.Node; kwargs...)

Make a diagram of the MTG tree, paired with a Makie.jl backend.

See also apply_recipe for the same plot with a Plots.jl backend.

The main attributes are:

color: the color of the nodes

colormap: the colormap used if the color uses an attribute. By default it uses viridis.

Must be a ColorScheme from ColorSchemes or a Symbol with its name.

Examples

using PlantGeom, GLMakie

file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
# file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","coffee.opf")

opf = read_opf(file)

diagram(opf)

# We can also color the 3d plot with several options:
# With one shared color:
diagram(opf, color = :red)

# Or colouring by opf attribute, *e.g.* using the nodes Z coordinates:
diagram(opf, color = :ZZ)

PlantGeom.extend_pos — Method

Add a new point after (x1,y1) using same direction and length relative to it

PlantGeom.get_attr_type — Method

Get the attributes types in Julia DataType.

PlantGeom.get_color — Method

get_color(colormap, var <: AbstractArray, range_var)
get_color(colormap, var, range_var)

Map value(s) to colors from a colormap based on a range of values

PlantGeom.get_colormap — Method

get_colormap(colormap)

Get the colormap as a ColorScheme if it is a named color or ColorScheme

PlantGeom.get_ref_mesh_index — Function

get_ref_mesh_index!(node, ref_meshes = get_ref_meshes(node))
get_ref_mesh_index(node, ref_meshes = get_ref_meshes(node))

Get the index of the reference mesh used in the current node.

Notes

Please use the ref_meshes argument preferably as not giving it make the function visit the root node each time otherwise, and it can become a limitation when traversing a big MTG.

PlantGeom.get_ref_mesh_index! — Function

get_ref_mesh_index!(node, ref_meshes = get_ref_meshes(node))
get_ref_mesh_index(node, ref_meshes = get_ref_meshes(node))

Get the index of the reference mesh used in the current node.

Notes

Please use the ref_meshes argument preferably as not giving it make the function visit the root node each time otherwise, and it can become a limitation when traversing a big MTG.

PlantGeom.get_ref_meshes — Method

get_ref_meshes(mtg)

Get all reference meshes from an mtg, usually from an OPF.

Examples

using PlantGeom
file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshes = get_ref_meshes(opf)

using GLMakie
viz(meshes)

PlantGeom.get_ref_meshes_color — Method

get_ref_meshes_color(meshes::RefMeshes)

Get the reference meshes colors (only the diffuse part for now).

Examples

using MultiScaleTreeGraph, PlantGeom
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshes = get_ref_meshes(opf)
PlantGeom.get_ref_meshes_color(meshes)

PlantGeom.map_coord — Method

map_coord(f, mesh, coord)

Apply function f over the mesh coordinates coord. Values for coord can be 1 for x, 2 for y and 3 for z.

PlantGeom.materialBDD_to_material — Method

Parse a material in opf format to a Phong material.

PlantGeom.meshBDD_to_meshes — Method

meshBDD_to_meshes(x)

Examples

using MultiScaleTreeGraph
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)
meshBDD_to_meshes(opf.attributes[:meshBDD])

PlantGeom.mtg_coordinates_df — Function

mtg_coordinates_df(mtg, attr; force = false)
mtg_coordinates_df!(mtg, attr; force = false)

Extract the coordinates of the nodes of the mtg and the coordinates of their parents (:XXfrom, :YYfrom, :ZZ_from) and output a DataFrame. Optionally you can also provide an attribute to add to the output DataFrame too by passing its name as a symbol to attr.

The coordinates are computed using coordinates! if missing, or if force = true.

PlantGeom.mtg_to_opf_link — Method

mtg_to_opf_link(link)

Takes an MTG link as input ("/", "<" or "+") and outputs its corresponding link as declared in the OPF format ("decomp", "follow" or "branch")

PlantGeom.mtg_topology_to_xml! — Function

mtg_topology_to_xml!(node, xml_parent)

Write the MTG topology, attributes and geometry into XML format. This function is used to write the "topology" section of the OPF.

PlantGeom.parse_geometry — Method

Parse the geometry element of the OPF.

Note

The transformation matrix is 3*4. elem = elem.content

PlantGeom.parse_materialBDD! — Method

Parse the materialBDD using parse_opf_elements!

PlantGeom.parse_meshBDD! — Method

Parse the meshBDD using parse_opf_array

PlantGeom.parse_opf_array — Function

Parse an array of values from the OPF into a Julia array (Arrays in OPFs are not following XML recommendations)

PlantGeom.parse_opf_attributeBDD! — Method

Parse the opf attributes as a Dict.

PlantGeom.parse_opf_elements! — Method

Generic parser for OPF elements.

Arguments

opf::OrderedDict: the opf Dict (using [XMLDict.xml_dict])

elem_types::Array: the target types of the element (e.g. "[String, Int64]")

Details

elem_types should be of the same length as the number of elements found in each item of the subchild. elem_types = [Float64, Float64, Float64, Float64, Float64, Float64]

PlantGeom.parse_opf_topology! — Function

Parser for OPF topology.

Note

The transformation matrices in geometry are 3*4. parseopftopology!(elem, nodei, features) node = elem mtg = nodei features = getattrtype(opf_attr[:attributeBDD])

Debugging:

mtg = nothing

node = elem mtg = nodei parseopftopology!( node, nothing, getattrtype(opfattr[:attributeBDD]), attrtype, mtgtype, ref_meshes )

PlantGeom.parse_ref_meshes — Method

parse_ref_meshes(mtg)

Parse the reference meshes of an OPF into RefMeshes.

PlantGeom.read_opf — Function

read_opf(file, attr_type = Dict, mtg_type = MutableNodeMTG)

Read an OPF file, and returns an MTG.

Arguments

file::String: The path to the opf file.

attr_type::DataType = Dict: the type used to hold the attribute values for each node.

mtg_type = MutableNodeMTG: the type used to hold the mtg encoding for each node (i.e.

link, symbol, index, scale). See details section below.

Details

attr_type should be:

NamedTuple if you don't plan to modify the attributes of the mtg, e.g. to use them for

plotting or computing statistics...

MutableNamedTuple if you plan to modify the attributes values but not adding new attributes

very often, e.g. recompute an attribute value...

Dict or similar (e.g. OrderedDict) if you plan to heavily modify the attributes, e.g.

adding/removing attributes a lot

The MultiScaleTreeGraph package provides two types for mtg_type, one immutable (MultiScaleTreeGraph.NodeMTG), and one mutable (MultiScaleTreeGraph.MutableNodeMTG). If you're planning on modifying the mtg encoding of some of your nodes, you should use MultiScaleTreeGraph.MutableNodeMTG, and if you don't want to modify anything, use MultiScaleTreeGraph.NodeMTG instead as it should be faster.

Note

See the documentation of the MTG format from the MTG package documentation for further details, e.g. The MTG concept.

Returns

The MTG root node.

Examples

using PlantGeom
file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
# file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","coffee.opf")
opf = read_opf(file)

PlantGeom.read_ply — Method

read_ply(file)

Read a ply file into a Meshes.SimpleMesh format. Code taken from Meshes.jl documentation.

Note: Remove when we have cylinder triangulation in Meshes.jl (and remove the dependency to PlyIO)

PlantGeom.refmesh_to_mesh — Function

refmesh_to_mesh!(node)
refmesh_to_mesh(node)

Compute a node mesh based on the reference mesh, the transformation matrix and the tapering. The mutating version adds the new mesh to the mesh field of the geometry attribute of the node.

Examples

using PlantGeom
file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)

node = opf[1][1][1]

new_mesh = refmesh_to_mesh(node)

using MeshViz, GLMakie
viz(new_mesh)

PlantGeom.refmesh_to_mesh! — Function

refmesh_to_mesh!(node)
refmesh_to_mesh(node)

Compute a node mesh based on the reference mesh, the transformation matrix and the tapering. The mutating version adds the new mesh to the mesh field of the geometry attribute of the node.

Examples

using PlantGeom
file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)

node = opf[1][1][1]

new_mesh = refmesh_to_mesh(node)

using MeshViz, GLMakie
viz(new_mesh)

PlantGeom.rotate_point — Method

Rotate a point (x1,y1) around (x0, y0) with angle.

PlantGeom.taper — Method

Returns a tapered mesh using dDwn and dUp based on the geometry of an input mesh. Tapering a mesh transforms it into a tapered version (i.e. pointy) or enlarged object, e.g. make a cone from a cylinder.

PlantGeom.transform_mesh! — Method

transform_mesh!(node::Node, transformation)

Add a new CoordinateTransformations.jl transformation to the node geometry transformation field. The transformation is composed with the previous transformation if any.

transformation must be a CoordinateTransformations.jl transformation.

It is also possible to invert a transformation using inv from CoordinateTransformations.jl.

Examples

using PlantGeom, MultiScaleTreeGraph, GLMakie, Rotations, CoordinateTransformations

file = joinpath(dirname(dirname(pathof(PlantGeom))), "test", "files", "simple_OPF_shapes.opf")
opf = read_opf(file)

# Visualize the mesh as is:
viz(opf)

# Copy the OPF, and translate the whole plant by 15 in the y direction (this is in cm, the mesh comes from XPlo):
opf2 = deepcopy(opf)
transform!(opf2, x -> transform_mesh!(x, Translation(0, 15, 0)))
viz!(opf2) # Visualize it again in the same figure

# Same but rotate the whole plant around the X axis:
opf3 = deepcopy(opf)
transform!(opf3, x -> transform_mesh!(x, LinearMap(RotX(0.3))))
# NB: we use Rotations.jl's RotX here. Input in radian, use rad2deg and deg2rad if needed.
viz!(opf3)

# Same but rotate only the second leaf around the Z axis:
opf4 = deepcopy(opf)
# Build the meshes from the reference meshes (need it because we want the coordinates of the parent):
transform!(opf4, refmesh_to_mesh!)

# Get the second leaf in the OPF:
leaf_node = get_node(opf4, 8)

# Get the parent node (internode) Z coordinates:
parent_zmax = zmax(leaf_node.parent)

# Define a rotation of the mesh around the Z axis defined by the parent node max Z:
transformation = recenter(LinearMap(RotZ(1.0)), Point3(0.0, 0.0, parent_zmax))

# Update the transformation matrix of the leaf and its mesh:
transform_mesh!(leaf_node, transformation)

# Plot the result:
viz(opf)
viz!(opf4)

PlantGeom.write_opf — Method

write_opf(opf, file)

Write an MTG with explicit geometry to disk as an OPF file.

Notes

Node attributes :mesh, :ref_meshes and :geometry are treated as reserved keywords and should not be used without knowing their meaning:

:ref_meshes: a RefMeshes structure that holds the MTG reference meshes.

:geometry: a Dict of four:

:shapeIndex: the index of the node reference mesh

:dUp: tappering in the upper direction

:dDwn: tappering in the bottom direction

:mat: the transformation matrix (4x4)

:mesh: a SimpleMesh computed from a reference mesh (:ref_meshes) and a transformation

matrix (:geometry).

Examples

using PlantGeom
file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
opf = read_opf(file)

write_opf("test.opf", opf)
file = "test.opf"
opf2 = read_opf(file)
viz(opf2)

PlantGeom.xmax — Function

xmax(x)
ymax(x)
zmax(x)

Get the maximum x, y or z coordinates of a mesh or a Node.

PlantGeom.xmin — Function

xmin(x)
ymin(x)
zmin(x)

Get the minimum x, y or z coordinates of a mesh or a Node.

PlantGeom.ymax — Function

xmax(x)
ymax(x)
zmax(x)

Get the maximum x, y or z coordinates of a mesh or a Node.

PlantGeom.ymin — Function

xmin(x)
ymin(x)
zmin(x)

Get the minimum x, y or z coordinates of a mesh or a Node.

PlantGeom.zmax — Function

xmax(x)
ymax(x)
zmax(x)

Get the maximum x, y or z coordinates of a mesh or a Node.

PlantGeom.zmin — Function

xmin(x)
ymin(x)
zmin(x)

Get the minimum x, y or z coordinates of a mesh or a Node.

RecipesBase.plot — Function

plot(opf::MultiScaleTreeGraph.Node; kwargs...)
plot!(opf::MultiScaleTreeGraph.Node; kwargs...)

Make a diagram of the MTG tree, paired with a Plots.jl backend.

See also diagram for the same plot with a Makie.jl backend.

Attributes

mode = "2d": The mode for plotting, either "2d" or "3d"

node_color = :black: the node color, can be a color or any MTG attribute

edge_color = node_color: same as node_color, but for the edges

colormap = :viridis: the colormap used for coloring

color_missing = RGBA(0, 0, 0, 0.3): The color used for missing values

Examples

# import Pkg; Pkg.add("PlotlyJS")
using Plots, PlantGeom
plotlyjs()

file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
# file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","coffee.opf")

opf = read_opf(file)

plot(opf, node_color = :Length)

RecipesBase.plot! — Function

plot(opf::MultiScaleTreeGraph.Node; kwargs...)
plot!(opf::MultiScaleTreeGraph.Node; kwargs...)

Make a diagram of the MTG tree, paired with a Plots.jl backend.

See also diagram for the same plot with a Makie.jl backend.

Attributes

mode = "2d": The mode for plotting, either "2d" or "3d"

node_color = :black: the node color, can be a color or any MTG attribute

edge_color = node_color: same as node_color, but for the edges

colormap = :viridis: the colormap used for coloring

color_missing = RGBA(0, 0, 0, 0.3): The color used for missing values

Examples

# import Pkg; Pkg.add("PlotlyJS")
using Plots, PlantGeom
plotlyjs()

file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","simple_OPF_shapes.opf")
# file = joinpath(dirname(dirname(pathof(PlantGeom))),"test","files","coffee.opf")

opf = read_opf(file)

plot(opf, node_color = :Length)